# perf(sensor): no alloc trace writer