--- a/sensor/debugger/k8s/trace_bench_test.go
+++ b/sensor/debugger/k8s/trace_bench_test.go
@@ -1,0 +1,21 @@
+package k8s
+
+import (
+	"path"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func BenchmarkWrite(b *testing.B) {
+	dir := b.TempDir()
+	data := []byte("abc")
+	writer, err := NewTraceWriter(path.Join(dir, "test"))
+	assert.NoError(b, err)
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		_, err := writer.Write(data)
+		assert.NoError(b, err)
+	}
+}