--- a/sensor/debugger/k8s/trace_test.go
+++ b/sensor/debugger/k8s/trace_test.go
@@ -1,0 +1,49 @@
+package k8s
+
+import (
+	"os"
+	"path"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func BenchmarkWrite(b *testing.B) {
+	dir := b.TempDir()
+	data := []byte("abc")
+	writer, err := NewTraceWriter(path.Join(dir, "test"))
+	assert.NoError(b, err)
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		_, err := writer.Write(data)
+		assert.NoError(b, err)
+	}
+}
+
+func TestWriter(t *testing.T) {
+	writer, err := NewTraceWriter("")
+	assert.Error(t, err)
+	assert.Nil(t, writer)
+
+	dir := t.TempDir()
+	filePath := path.Join(dir, "test")
+
+	writer, err = NewTraceWriter(filePath)
+	n, err := writer.Write([]byte("abc"))
+	assert.NoError(t, err)
+	assert.Equal(t, 4, n)
+
+	n, err = writer.Write([]byte("1337"))
+	assert.NoError(t, err)
+	assert.Equal(t, 5, n)
+
+	err = writer.Close()
+	assert.NoError(t, err)
+	err = writer.Close()
+	assert.Error(t, err)
+
+	data, err := os.ReadFile(filePath)
+	assert.NoError(t, err)
+	assert.Equal(t, data, []byte("abc\n1337\n"))
+}