--- a/tools/local-sensor/main.go
+++ b/tools/local-sensor/main.go
@@ -331,10 +331,8 @@ func main() {
 	}
 
 	if localConfig.RecordK8sEnabled {
-		traceRec := &k8s.TraceWriter{
-			Destination: path.Clean(localConfig.RecordK8sFile),
-		}
-		if err := traceRec.Init(); err != nil {
+		traceRec, err := k8s.NewTraceWriter(path.Clean(localConfig.RecordK8sFile))
+		if err != nil {
 			log.Fatalln(err)
 		}
 		sensorConfig.WithTraceWriter(traceRec)