### Description

This PR changes the implementation of sensor trace writer to not open file on every write and not append delimiter to bytes to write. This will reduce allocations as append may result in doubling bytes slice and opening a file result with additional 2 syscalls to open and close it. Instead we use sync to flush data to drive.

```python
benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/stackrox/rox/sensor/debugger/k8s
        │   old.txt    │               new.txt               │
        │    sec/op    │   sec/op     vs base                │
Write     11.312µ ± 3%   3.766µ ± 4%  -66.71% (p=0.000 n=10)

        │  old.txt   │              new.txt              │
        │    B/op    │   B/op    vs base                 │
Write     160.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

        │  old.txt   │               new.txt               │
        │ allocs/op  │ allocs/op   vs base                 │
Write     4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```

- [x] CHANGELOG update is not needed
- [x] Documentation is not needed
### Testing
- [x] inspected CI results
#### Automated testing
- [x] modified existing tests
- [x] contributed **no automated tests**
#### How I validated my change
CI